Moving consts and groupByArray to /scripts/utils files
